separate compontents to reduce the code and scale easier
