+++ package.json
@@ -27,6 +27,7 @@
     "@ionic-native/status-bar": "^4.5.3",
     "@ionic/storage": "^2.1.3",
     "angular2-moment": "^1.8.0",
+    "closure-webpack-plugin": "0.0.23",
     "cordova-android": "7.0.0",
     "cordova-ios": "4.5.4",
     "cordova-plugin-device": "^2.0.1",
