+++ src/components/post-render/post-render.scss
@@ -1,63 +1,3 @@
 post-render {
 
-    #postImage { 
-        height: 280px; 
-        width: 100%; 
-        overflow: hidden; 
-    }
-    .info {
-        color: gray;
-    }
-    ion-icon {
-        color: gray;
-    }
-    .mt3{
-        margin-top: 3px;
-      }
-
-    .img {
-        position: relative;
-        width:  100% !important;
-        background-position: 50% 50%;
-        background-repeat:   no-repeat;
-    }
-
-    .voters_image {
-        display:inline-block;
-        width: 25px; 
-        height: 25px; 
-        
-        img {
-            border-radius: 50%;
-            border: 1.5px solid white;
-        }
-    }
-
-    .voters_image:not(:first-of-type) {
-        margin-left: -6px;
-    }
-
-    .voters {
-        float: left;
-        padding: 5px;
-    }
-
-    .likes {
-        float: left;
-        padding: 10px 10px 10px 0;
-
-        p {
-            font-weight: 500;
-            font-size: 12px;
-        }
-    }
-
-    .button-clear {
-        color: #565656 !important;
-    }
-
-    ion-spinner {
-        top: 0 !important;
-    }
-
 }
