+++ src/pages/TabsSections/feed/feed.ts
@@ -5,6 +5,7 @@ import { SteemProvider } from '../../../providers/steem/steem';
 import { SteemConnectProvider } from 'providers/steemconnect/steemconnect';
 import { Observable } from 'rxjs/Observable';
 import { feedTemplate } from './feed.template';
+import { ImageLoaderConfig } from 'ionic-image-loader';
 
 @Component({
   selector: 'section-scss',
@@ -22,7 +23,14 @@ export class FeedPage {
               private appCtrl: App,
               private steemConnect: SteemConnectProvider,
               private zone: NgZone,
-              private cdr: ChangeDetectorRef) {
+              private cdr: ChangeDetectorRef,
+              private imageLoaderConfig: ImageLoaderConfig) {
+
+                this.imageLoaderConfig.setBackgroundSize('cover');
+    this.imageLoaderConfig.setHeight('200px');
+    this.imageLoaderConfig.setFallbackUrl('assets/placeholder2.png');
+    this.imageLoaderConfig.setImageReturnType('base64');
+    this.imageLoaderConfig.enableFallbackAsPlaceholder(true);
 
     /**
      * Subscribe to the user object in the auth provider
