+++ src/pages/TabsSections/sections.scss
@@ -40,4 +40,6 @@ section-scss {
     .noborder {
       border: none !important;
     }
+
+    
 }
