+++ src/pages/post-single/post-single.ts
@@ -3,7 +3,6 @@ import { IonicPage, NavController, NavParams } from 'ionic-angular';
 import { Post } from 'models/models';
 import marked from 'marked';
 import { SteemProvider } from '../../providers/steem/steem';
-import { DataProvider } from '../../providers/data/data';
 
 @IonicPage({
   priority: 'high'
@@ -21,8 +20,7 @@ export class PostSinglePage {
 
   constructor(public navCtrl: NavController, 
               public navParams: NavParams,
-              public steemData: SteemProvider,
-              public dataProvider: DataProvider,) {
+              public steemData: SteemProvider) {
 
     this.post = this.navParams.get('post');
     
@@ -38,14 +36,7 @@ export class PostSinglePage {
     //this.body = md.render(this.post.body)
     //this.body = this.post.body.replace(image_url,this.embedService.embed('$1'))
     this.steemData.getComments({author: this.post.author, permlink: this.post.permlink}).subscribe(data => {
-      console.log(data)
-    })
-
-    this.dataProvider.getContent([this.post.author], [this.post.permlink]).subscribe(data => {
-      console.log(data);
-      this.meta = JSON.parse(data.json_metadata);
-      this.tags = this.meta.tags;
-    //  console.log(this.tags);
+      //console.log(data)
     })
 
   }
